refactor(Segments): use correct naming scheme
